requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
tiktoken==0.4.0
python-dotenv==1.0.0
pytest==7.3.1
httpx==0.24.0
slowapi==0.1.7
pydantic==1.10.7
streamlit==1.22.0
requests==2.28.2

# Generated by Copilot
